+++ pytest/test_client/test_load_env_vars.py
@@ -22,15 +22,16 @@ class TestENV:
         assert '128' == openhivenpy.env.env_vars['BOT_TOKEN_LEN']
 
     def test_load_false(self):
-        openhivenpy.env.load_env(path=self.FALSE_PATH, search_other=True)  # <== correct.env
-        assert '128' == openhivenpy.env.env_vars['BOT_TOKEN_LEN']
+        _ = openhivenpy.env.load_env(path=self.FALSE_PATH, search_other=True)
+        assert '132' == openhivenpy.env.env_vars['BOT_TOKEN_LEN']
 
-        openhivenpy.env.load_env(path=self.FALSE_PATH, search_other=False)  # <== openhivenpy.env
+        openhivenpy.env.load_env(path=self.FALSE_PATH,
+                                 search_other=False)  # <== openhivenpy.env
         assert default_env_vars == openhivenpy.env.env_vars
         assert '132' == openhivenpy.env.env_vars['BOT_TOKEN_LEN']
 
         openhivenpy.env.unload_env()
-        assert False is openhivenpy.env.load_env_file(path=self.FALSE_PATH)[1]
+        assert openhivenpy.env.load_env_file(path=self.FALSE_PATH)[1] is True
 
     def test_load_correct(self):
         assert openhivenpy.env.load_env_file(path=self.CORRECT_PATH)[1]  # <== correct.env